+++ packages/log-http/defaults.js
@@ -6,6 +6,9 @@ const req = require('./req');
 const res = require('./res');
 
 module.exports = Object.assign({}, defaults, {
+    customLogLevel(response) {
+        return response.statusCode >= 500 ? 'error' : 'info';
+    },
     serializers: {
         err,
         req,